requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.0
python-socketio==5.9.0
instagrapi==2.0.0
tweepy==4.14.0
cryptography==41.0.0
pyyaml==6.0.1
jsonschema==4.19.0
watchdog==3.0.0
SQLAlchemy==2.0.49
pytest==7.4.0
pytest-cov==4.1.0
python-dotenv==1.0.0
Flask-Login==0.6.3
